requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.3.0
gunicorn==21.2.0
whitenoise==6.6.0